# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/bankingapp-0.0.1-SNAPSHOT",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/user/account/{accountId}/transactions": {
      "get": {
        "tags": [
          "Transaction "
        ],
        "operationId": "accountTransaction",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTransactionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transaction "
        ],
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTransactionDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/user/{userId}/account/transactions": {
      "get": {
        "tags": [
          "Transaction "
        ],
        "operationId": "getTransactionsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTransactionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions": {
      "get": {
        "tags": [
          "Transaction "
        ],
        "operationId": "getAllTransactions",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTransactionDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TransactionDTO": {
        "required": [
          "currency",
          "transactionAmount",
          "transactionType"
        ],
        "type": "object",
        "properties": {
          "transactionAmount": {
            "type": "integer",
            "format": "int32"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "CREDIT",
              "DEBIT"
            ]
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string"
          }
        }
      },
      "ResponseTransactionDTO": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "transactionAmount": {
            "type": "integer",
            "format": "int32"
          },
          "transactionMessage": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "CREDIT",
              "DEBIT"
            ]
          },
          "transactionStatus": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "PENDING",
              "FAILED"
            ]
          }
        }
      },
...
    }
  }
}